* Cache field code lookups. * /META support. * Added get_string() and get_const(), prevent calling getdata for non-vectors. * Numerous compiler warnings fixed. * Fixed get_error_string() for GCC 4.3 C99 compliance.

+++ test/legacy_nframes.c
@@ -19,7 +19,7 @@ int main(void)
   const char* data = __TEST__ "dirfile/data";
   const char* format_data = "data RAW UINT8 1\n";
   int fd, error;
-  const int len = strlen(data);
+  const size_t len = strlen(data);
 
   mkdir(filedir, 0777);
